requirements.txt
selenium
bs4
python-dotenv
bcrypt
snowflake-sqlalchemy
pandas
pydantic
openai
fastapi
streamlit
langchain
tiktoken
langchain-openai
matplotlib